+++ CHANGELOG.md
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SharePointDsc
   - Updated all resources and Invoke-SPDscCommand function to automatically
     load Utils module, which broke with the new CI
+  - Extended Convert-SPDscHashtableToString function to support complex types
+    in arrays and the CIMInstance type
 - SPConfigWizard
   - Updated checks in Set method to make sure the resource also runs when
     a language pack is installed